+++ my_pgp
@@ -2,7 +2,7 @@
 
 from Parser import *
 
-from AES import Aes
+from Aes.AES import Aes
 from XOR import Xor
 from RSA import Rsa
 from PGP_XOR import PgpXor
